% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neigh_ob.R
\name{neigh_ob}
\alias{neigh_ob}
\title{a spatial adjecency data for the polygon grid 'shp'}
\format{an object of class 'nb'}
\description{
A spatial adjecency object of class 'nb' for the polygon shapefile 'shp'.
See documentation for \code{\link[spdep]{card}} for details on
structure of the object.
}

