% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangemodelR-package.R
\docType{package}
\name{rangemodelR-package}
\alias{rangemodelR}
\alias{rangemodelR-package}
\title{Mid-Domain Effect and Species Richness Patterns}
\description{
Range Cohesion and Range Scatter Models in R
}
\details{
\tabular{ll}{
Package: \tab rangemodelR\cr
Type: \tab Package\cr
Version: \tab 1.0.6\cr
Date: \tab 2023-06-08\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab No\cr
}

Used for generating randomized community matrices under
strict range cohesion. The package can handle data where species occurrence
are recorded across one or two dimensions.One dimensional data are typically
for sites ordered along a variable of interest such
as elevation or latitude.\code{\link{rangemod1d}} is used for such data.
Range cohesion is enforced using an 'nb' object
(generated by \code{\link[spdep]{poly2nb}}
of the \pkg{\link[spdep]{spdep}} package).
}
\references{
\itemize{
                     \item Rahbek, C., Gotelli, N., Colwell, R., Entsminger, G., Rangel, T. & Graves, G. (2007)
             Predicting continental-scale patterns of bird species richness with spatially explicit
             models. Proceedings of the Royal Society B: Biological Sciences, 274, 165.

                     \item Colwell, Robert K., and David C. Lees. "The mid-domain effect:
           geometric constraints on the geography of species richness."
           Trends in ecology & evolution 15.2 (2000): 70-76.
           }
}
\author{
Aniruddha Marathe \email{aniruddha.pravin.marathe@gmail.com}
}
\keyword{internal}
\keyword{package}
