% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spmat.R
\name{spmat}
\alias{spmat}
\title{Site by Species Matrix of Hypothetical Assemblage of 20 Species Across the
Grid 'shp'}
\format{a dataframe with 100 rows and 20 columns}
\description{
A site by species matrix with cells of the grid in shapefile \link{shp}
in rows and 20 species in columns. Presence of species is denoted
 by '1' and absence by '0'.
}

