% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rangemodelR-package.R
\docType{package}
\name{rangemodelR-package}
\alias{rangemodelR}
\alias{rangemodelR-package}
\title{impliments range cohesion models}
\description{
impliments range scatter and range cohesion models
}
\details{
\tabular{ll}{
Package: \tab rangemodelR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-08-09\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

The package is usefull for generating randomized community matrices under
strict range cohesion. The package can handle data where species occurence
are recorded across one or two dimentions.One dimentional data are typically
for sites ordered along a variable of interest such
as elevation or latitude.\code{\link{rangemod.1d}} is used for such data. Two
dimentional data have species occurences recorded on grids with known
geographic coordinates and are handled with function \code{\link{rangemod.2d}}.
Range cohesion is enforced using an 'nb' object
(generated by \code{\link[spdep]{poly2nb}}
of the \pkg{\link[spdep]{spdep}} package).
}
\author{
Aniruddha Marathe \email{aniruddha.pravin.marathe@gmail.com}
}
\references{
Rahbek, C., Gotelli, N., Colwell, R., Entsminger, G., Rangel, T. & Graves, G. (2007)
             Predicting continental-scale patterns of bird species richness with spatially explicit
             models. Proceedings of the Royal Society B: Biological Sciences, 274, 165.
}
\keyword{package}

