% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp.R
\name{shp}
\alias{shp}
\title{Polygon Grid Roughly Located on Central India}
\format{an ESRI shapefile imported with \code{\link[maptools]{readShapeSpatial}} with
       four columns in attribute table}
\description{
A grid with 100 cells, each \eqn{1^0} by \eqn{1^0} in size roughly located on
central India. Each cell is assigned a hypothetical assemblage of varying
species richness from a hypothetical regional pool of 20 species.
The attribute table of the shape file has one uniqe id field observed species
richness of each cell and one expanatory variable. Observed species richness
is generated by assigning a randomly selected number of species among the 20
species in the regional pool.
\itemize{
   \item ID  a unique id field
   \item obs_rich  observed species richness
   \item trend  a variable with spatial trend obtained using a linear function
               of coordinates}
}
